Refac: 기수 관련 코드 분리
