+++ src/main/java/com/gdsc_knu/official_homepage/service/admin/AdminApplicationService.java
@@ -127,50 +127,4 @@ public class AdminApplicationService {
                 = applicationRepository.findByNameContaining(PageRequest.of(page,size), name);
         return PagingResponse.from(applicationPage, AdminApplicationResponse.Overview::from);
     }
-
-    @Transactional(readOnly = true)
-    public List<AdminApplicationResponse.ClassYearResponse> getClassYearList() {
-        return classYearRepository.findAll().stream().map(AdminApplicationResponse.ClassYearResponse::from).toList();
-    }
-
-    @Transactional(readOnly = true)
-    public AdminApplicationResponse.ClassYearResponse getClassYear(Long id) {
-        ClassYear classYear = classYearRepository.findById(id)
-                .orElseThrow(() -> new CustomException(ErrorCode.CLASS_YEAR_NOT_FOUND));
-        return AdminApplicationResponse.ClassYearResponse.from(classYear);
-    }
-
-    @Transactional
-    public void addClassYear(AdminApplicationRequest.ClassYearRequest request) {
-        validateClassYear(request);
-        classYearRepository.save(request.toEntity());
-    }
-
-    @Transactional
-    public void updateClassYear(Long id, AdminApplicationRequest.ClassYearRequest request) {
-        ClassYear classYear = classYearRepository.findById(id)
-                .orElseThrow(() -> new CustomException(ErrorCode.CLASS_YEAR_NOT_FOUND));
-        validateClassYear(request);
-        classYear.update(request.getName(), request.getApplyStartDateTime(), request.getApplyEndDateTime());
-        classYearRepository.save(request.toEntity());
-    }
-
-    @Transactional
-    public void deleteClassYear(Long id) {
-        try {
-            classYearRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new CustomException(ErrorCode.CLASS_YEAR_NOT_FOUND);
-        }
-    }
-
-    private void validateClassYear(AdminApplicationRequest.ClassYearRequest request) {
-        classYearRepository.findByName(request.getName())
-                .ifPresent(classYear -> {
-                    throw new CustomException(ErrorCode.CLASS_YEAR_DUPLICATED);
-                });
-        if (request.getApplyStartDateTime().isAfter(request.getApplyEndDateTime()) || request.getApplyEndDateTime().isBefore(request.getApplyStartDateTime())) {
-            throw new CustomException(ErrorCode.INVALID_CLASS_YEAR);
-        }
-    }
 }
